% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{counter}
\alias{counter}
\title{Title counter}
\usage{
counter(dat, var)
}
\arguments{
\item{dat}{A dataset}

\item{var}{A categorical variable}
}
\value{
A small dataset that contains counts of each category
}
\description{
Title counter
}
\examples{

counter(infert, "education")

}
