% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumuisha.R
\name{jumuisha}
\alias{jumuisha}
\title{Title jumuisha}
\usage{
jumuisha(x, y)
}
\arguments{
\item{x}{First number}

\item{y}{Second number}
}
\value{
sum of x and y
}
\description{
Description: Takes in two numbers and adds them
}
\examples{

jumuisha(5, 7)

}
