% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salamu.R
\name{salamu}
\alias{salamu}
\title{Title salamu}
\usage{
salamu(name)
}
\arguments{
\item{name}{Name}
}
\value{
A message
}
\description{
Description Generates a message given someone's name
}
\examples{
salamu("Shel")

}
